Tighten devtools causal architecture
